cleanup, fix delay bug
