+++ src/station.c
@@ -12,20 +12,7 @@
 #include "graph.h"
 #include "station.h"
 #include "track.h"
-
-//Helper Functions
-void SWAP (int *a, int *b) {
-    int tmp = *a;
-    *a = *b;
-    *b = tmp;
-}
-void SWAP_SHORT (short *a, short *b) {
-    short tmp = *a;
-    *a = *b;
-    *b = tmp;
-}
-
-
+#include "utils.h"
 
 
 //Init function
@@ -144,6 +131,7 @@ void station_event (station_state *s, tw_bf *bf, message *in_msg, tw_lp *lp) {
                 // All these passengers got on here I guess
                 msg->source = self;
                 msg->next_arrival = curr_track->next_arrival;
+                msg->delayed=0; //NOTE: this might be unnecessary?
                 //tw_output(lp, "[%.3f] ST %d: Sending ack message to %d on track %d!\n", tw_now(lp), self, in_msg->source, curr_track->track_id);
 #if DEBUG_FILE_OUTPUT
                 fprintf(node_out_file, "[ST %d]: Sending ack message to %lu on track %d!\n", self, in_msg->source, curr_track->track_id);
