+++ src/utils.h
@@ -1,0 +1,12 @@
+// utils.h
+// stupid helper functions that both stations and trains need
+
+#ifndef _utils_h
+#define _utils_h
+
+// Some handy utils
+void SWAP (int *a, int *b);
+void SWAP_UL (unsigned long *a, unsigned long *b);
+void SWAP_SHORT (short *a, short *b);
+
+#endif
